Disable jaeger for now - Move open telemetry event listener to another file
